feat(tests): adding edit an application test
