+++ tests/sanity/tests/model/recruiting/applications-details-page.ts
@@ -1,77 +1,11 @@
 import { type Locator, type Page } from '@playwright/test'
-import { NewApplication, TalentName } from './types'
-import { CommonPage } from '../common-page'
-import { generateId } from '../../utils'
 
-export class ApplicationsPage extends CommonPage {
+export class ApplicationsDetailsPage {
   readonly page: Page
-  readonly pageHeader: Locator
-  readonly buttonCreateApplication: Locator
-  readonly buttonTalentSelector: Locator
-  readonly buttonSpaceSelector: Locator
-  readonly buttonAssignedRecruiter: Locator
-  readonly buttonCreateNewApplication: Locator
+  readonly inputCommentMessage: Locator
 
   constructor (page: Page) {
-    super()
     this.page = page
-    this.pageHeader = page.locator('span[class*="header"]', { hasText: 'Applications' })
-    this.buttonCreateApplication = page.locator('button > span', { hasText: 'Application' })
-    this.buttonTalentSelector = page.locator('div[id="vacancy.talant.selector"]')
-    this.buttonSpaceSelector = page.locator('div[id="space.selector"]')
-    this.buttonAssignedRecruiter = page.locator('button div.label', { hasText: 'Assigned recruiter' })
-    this.buttonCreateNewApplication = page.locator('form[id="recruit:string:CreateApplication"] button[type="submit"]')
-  }
-
-  async createNewApplication (data: NewApplication): Promise<void> {
-    await this.buttonCreateApplication.click()
-
-    await this.selectTalent(data.talentsName != null ? data.talentsName : 'first')
-    await this.selectVacancy(data.vacancy)
-    await this.selectRecruiter(data.recruiterName)
-
-    await this.buttonCreateNewApplication.click()
-  }
-
-  async createNewApplicationWithNewTalent (data: NewApplication): Promise<TalentName> {
-    const talentName: TalentName = {
-      firstName: `TestFirst-${generateId(4)}`,
-      lastName: `TestLast-${generateId(4)}`
-    }
-
-    await this.buttonCreateApplication.click()
-
-    await this.buttonTalentSelector.click()
-    await this.pressCreateButtonSelectPopup(this.page)
-    await this.createNewTalentPopup(this.page, talentName.firstName, talentName.lastName)
-
-    await this.selectVacancy(data.vacancy)
-    await this.selectRecruiter(data.recruiterName)
-
-    await this.buttonCreateNewApplication.click()
-
-    return talentName
-  }
-
-  async selectTalent (name: string): Promise<void> {
-    await this.buttonTalentSelector.click()
-    await this.fillSelectPopup(this.page, name)
-  }
-
-  async selectVacancy (name: string): Promise<void> {
-    await this.buttonSpaceSelector.click()
-    await this.fillSelectPopup(this.page, name)
-  }
-
-  async selectRecruiter (name: string): Promise<void> {
-    await this.buttonAssignedRecruiter.click()
-    await this.fillSelectPopup(this.page, name)
-  }
-
-  async openApplicationByTalentName (data: TalentName): Promise<void> {
-    await this.page.locator('span.ap-label', { hasText: `${data.lastName} ${data.firstName}` })
-      .locator('xpath=../../../../..')
-      .locator('div[class*="firstCell"]')
-      .click()
+    this.inputCommentMessage = page.locator('div.tiptap')
   }
 }
