+++ tests/sanity/tests/model/recruiting/types.ts
@@ -1,0 +1,10 @@
+export interface NewApplication {
+  talentsName?: string
+  vacancy: string
+  recruiterName: string
+}
+
+export interface TalentName {
+  firstName: string
+  lastName: string
+}
